Cleaned up templates and unified datetime displays.
