+++ src/hi/apps/attribute/templates/attribute/components/regular_attribute_card.html
@@ -1,5 +1,5 @@
 {% load icons %}
-{% load common_template_helpers %}
+{% load common_tags %}
 {% load attribute_extras %}
 {% comment %}
 V2 Attribute Card Component
