+++ src/hi/apps/attribute/templates/attribute/components/value_types/attribute_value_text.html
@@ -1,4 +1,4 @@
-{% load common_template_helpers %}
+{% load common_tags %}
 {% comment %}
 Text Attribute Value Component
 Renders textarea with special handling (hidden input + display textarea) for text attributes
