+++ src/hi/apps/console/templates/console/panes/entity_video_sensor_history.html
@@ -1,3 +1,4 @@
+{% load common_tags %}
 {% load video_tags %}
 {% load tz %}
 {% load icons %}
