+++ src/hi/apps/entity/templates/entity/modals/entity_state_history.html
@@ -1,5 +1,5 @@
 {% extends "modals/action_3.html" %}
-{% load common_template_helpers %}
+{% load common_tags %}
 
 {% block title_text %}
 {{ entity.name }}: History
