+++ src/hi/apps/location/templates/location/panes/location_view.html
@@ -1,5 +1,5 @@
 {% load static %}
-{% load common_template_helpers %}
+{% load common_tags %}
 <link rel="stylesheet" href="{% static location_view.svg_style_name.css_static_file_name %}">
 <div id="hi-location-view-main" class="hi-location-view">
   <svg id="{{ location_view.html_id }}"
