+++ src/hi/apps/sense/templates/sense/panes/sensor_response.html
@@ -1,4 +1,4 @@
-{% load common_template_helpers %}
+{% load common_tags %}
 <div class="{{ sensor_response.css_class }}">
   {% include_with_fallback sensor_response.sensor.entity_state.entity_state_type.value_template_name "sense/panes/sensor_value_default.html" %}
 </div>
