+++ src/hi/apps/user/templates/user/emails/signin_magic_link_message.html
@@ -1,5 +1,5 @@
 {% extends "notify/emails/base_email_message.html" %}
-{% load common_template_helpers %}
+{% load common_tags %}
 {% block title %}Sign In Access Code{% endblock %}
 {% block content %}
 <p>
